@@ -1,17 +1,17 @@
 name: dart_pad
-author: Dart Team <misc@dartlang.org>
 description: The UI client for a web based interactive Dart service.
 homepage: https://github.com/dart-lang/dart-pad
 environment:
-  sdk: '>=2.10.4 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   codemirror: ^0.6.0-0
+  collection: ^1.15.0
   fluttering_phrases: ^0.3.1
-  http: '>=0.12.2 <0.13.0'
-  json_annotation: ^3.1.0
-  logging: '>=0.11.4 <0.12.0'
-  markdown: ^3.0.0
+  http: ^0.13.0
+  json_annotation: ^4.0.0
+  logging: ^1.0.0
+  markdown: ^4.0.0
   meta: ^1.2.4
   protobuf: ^2.0.0
   split:
@@ -19,20 +19,26 @@ dependencies:
   html_unescape: ^2.0.0
   sass_builder: ^2.1.0
   mdc_web: ^0.6.0
-  yaml: ^2.2.0
-  checked_yaml: ^1.0.0
+  yaml: ^3.0.0
+  checked_yaml: ^2.0.0
   js: ^0.6.0
-  stream_transform: ^1.2.0
+  stream_transform: ^2.0.0
 
 dev_dependencies:
-  build_runner: any
-  build_test: any
-  build_web_compilers: any
-  collection: ^1.14.10
+  build_runner: ^2.0.0
+  build_test: ^2.0.0
+  build_web_compilers: ^3.0.0
   git: ^2.0.0
-  grinder: ^0.8.0
-  json_serializable: ^3.5.0
-  pedantic: ^1.11.0
-  shelf: ^0.7.5
-  shelf_static: ^0.2.8
+  grinder: ^0.9.0
+  json_serializable: ^4.0.0
+  lints: ^1.0.0
+  shelf: ^1.0.0
+  shelf_static: ^1.0.0
   test: ^1.15.7
+
+# waiting for the next material-components-web release.
+# Once released, it will need to be rolled into package:mdc_web.
+#
+# https://github.com/material-components/material-components-web/pull/7158
+dependency_overrides:
+  sass: 1.32.10
