@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-
 import 'github.dart';
 import 'web_server.dart';
 import 'workshop.dart';
@@ -8,10 +6,10 @@ abstract class WorkshopFetcher {
   Future<Workshop> fetch();
 
   factory WorkshopFetcher.github({
-    @required String owner,
-    @required String repo,
-    String ref,
-    String path,
+    required String owner,
+    required String repo,
+    String? ref,
+    String? path,
   }) =>
       GithubWorkshopFetcher(owner: owner, repo: repo, path: path, ref: ref);
 
