@@ -1,6 +1,5 @@
 import 'package:dart_pad/util/github.dart';
 import 'package:http/http.dart' as http;
-import 'package:meta/meta.dart';
 
 import 'exceptions.dart';
 import 'fetcher_impl.dart';
@@ -10,12 +9,12 @@ class GithubWorkshopFetcher extends WorkshopFetcherImpl {
 
   final String owner;
   final String repo;
-  final String ref;
-  final String path;
+  final String? ref;
+  final String? path;
 
   GithubWorkshopFetcher({
-    @required this.owner,
-    @required this.repo,
+    required this.owner,
+    required this.repo,
     this.ref,
     this.path,
   });
@@ -43,7 +42,13 @@ class GithubWorkshopFetcher extends WorkshopFetcherImpl {
     return Uri(
       scheme: 'https',
       host: _apiHostname,
-      pathSegments: ['repos', owner, repo, 'contents', ...filePath],
+      pathSegments: [
+        'repos',
+        owner,
+        repo,
+        'contents',
+        ...filePath as Iterable<String>
+      ],
       queryParameters: {if (ref != null) 'ref': ref},
     );
   }
