@@ -9,14 +9,14 @@ class EditorDocumentProperty implements Property<String> {
   final Document document;
   final String debugName;
 
-  EditorDocumentProperty(this.document, [this.debugName]);
+  EditorDocumentProperty(this.document, this.debugName);
 
   @override
   String get() => document.value;
 
   @override
   void set(String str) {
-    document.value = str ?? '';
+    document.value = str;
   }
 
   @override
