@@ -12,14 +12,13 @@ Meta _$MetaFromJson(Map json) {
         allowedKeys: const ['name', 'type', 'steps'],
         requiredKeys: const ['name', 'steps']);
     final val = Meta(
-      $checkedConvert(json, 'name', (v) => v as String),
+      $checkedConvert(json, 'name', (v) => v as String?),
       $checkedConvert(
           json,
           'steps',
-          (v) => (v as List)
-              ?.map((e) =>
-                  e == null ? null : StepConfiguration.fromJson(e as Map))
-              ?.toList()),
+          (v) => (v as List<dynamic>?)
+              ?.map((e) => StepConfiguration.fromJson(e as Map))
+              .toList()),
       type: $checkedConvert(json, 'type',
               (v) => _$enumDecodeNullable(_$WorkshopTypeEnumMap, v)) ??
           WorkshopType.dart,
@@ -34,36 +33,41 @@ Map<String, dynamic> _$MetaToJson(Meta instance) => <String, dynamic>{
       'steps': instance.steps,
     };
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
+K _$enumDecode<K, V>(
+  Map<K, V> enumValues,
+  Object? source, {
+  K? unknownValue,
 }) {
   if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
+    throw ArgumentError(
+      'A value must be provided. Supported values: '
+      '${enumValues.values.join(', ')}',
+    );
   }
 
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
+  return enumValues.entries.singleWhere(
+    (e) => e.value == source,
+    orElse: () {
+      if (unknownValue == null) {
+        throw ArgumentError(
+          '`$source` is not one of the supported values: '
+          '${enumValues.values.join(', ')}',
+        );
+      }
+      return MapEntry(unknownValue, enumValues.values.first);
+    },
+  ).key;
 }
 
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
+K? _$enumDecodeNullable<K, V>(
+  Map<K, V> enumValues,
   dynamic source, {
-  T unknownValue,
+  K? unknownValue,
 }) {
   if (source == null) {
     return null;
   }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
 }
 
 const _$WorkshopTypeEnumMap = {
